% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable1.R
\docType{data}
\name{enable1}
\alias{enable1}
\title{The enable1 word list}
\format{
An object of class \code{character} of length 172823.
}
\source{
\url{https://raw.githubusercontent.com/dolph/dictionary/master/enable1.txt}
}
\usage{
enable1
}
\description{
The enable1 word list
}
\keyword{datasets}
