% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/part1.R
\name{cpp_funnel}
\alias{cpp_funnel}
\alias{funnel}
\alias{r_funnel}
\title{Part 1 Word Funnel}
\usage{
cpp_funnel(a, b)

r_funnel(a, b)
}
\arguments{
\item{a}{The base word}

\item{b}{The candidate word}
}
\value{
A logical indicating if \code{a} can go into \code{b}
}
\description{
Part 1 Word Funnel
}
