% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/part2.R
\name{cpp_funnel2}
\alias{cpp_funnel2}
\alias{funnel2}
\alias{r_funnel2}
\title{Part 2 Word Funnel}
\usage{
cpp_funnel2(x, wordset)

r_funnel2(x, search_func)
}
\arguments{
\item{x}{The word to test.}

\item{wordset}{The enable1 list as an external pointer to a C++ unordered set.}

\item{search_func}{A function to check to see if a word is valid.}
}
\description{
Find the maximum depth of a word funnel.
}
