% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_set}
\alias{initialize_set}
\title{Create an External Pointer to a C++ Unordered Set}
\usage{
initialize_set(s)
}
\arguments{
\item{s}{The character vector to be used in the set.}
}
\description{
Create an External Pointer to a C++ Unordered Set
}
