% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1.R
\name{pt1_bonus}
\alias{pt1_bonus}
\title{Part 1 Bonus}
\usage{
pt1_bonus(x, search_func)
}
\arguments{
\item{x}{The word used to create funnel words.}

\item{search_func}{A function to check to see if a candidate word is in the
a dictionary.}
}
\description{
Create a list of all funnel words that exist in a dictionary.
}
\examples{
\dontrun{
data(enable1)
set_search_function <- set_search(enable1)
pt1_bonus("dragoon", set_search_function)
}
}
\seealso{
\link{search_func}
}
