% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1.R
\name{pt1_bonus2}
\alias{pt1_bonus2}
\title{Part 1 Bonus 2}
\usage{
pt1_bonus2(table, search_func)
}
\arguments{
\item{table}{A character vector of dictionary words. (I. e., \code{enable1})}

\item{search_func}{A function to check to see if a candidate word is in the
a dictionary.}
}
\description{
Find all words in the \code{enable1} dictionary that contains 5 legal funnel words.
}
