% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2.R
\name{pt2_bonus}
\alias{pt2_bonus}
\title{Part 2 Bonus}
\usage{
pt2_bonus(words, wordset)
}
\arguments{
\item{words}{The enable1 list as a character vector.}

\item{wordset}{The enable1 list as an external pointer to a C++ unordered
set.}
}
\description{
Find the word that has a funnel depth of 10
}
