% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_set}
\alias{initialize_set}
\title{Create an external pointer to a C++ unordered set}
\usage{
initialize_set(s)
}
\arguments{
\item{s}{The character vector to be used in the set.}
}
\description{
Create an external pointer to a C++ unordered set
}
